refactor(cassandra): spit tests into mutiple files
Improve code readability.
